chore: created tooltip component
